backend/foodgram/requirements.txt
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
djoser==2.1.0
drf-extra-fields==3.4.0
flake8==5.0.4
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytest==7.1.2
python-dotenv==0.21.0
reportlab==3.6.3
requests==2.22.0